WIP: Preprocessor decoupling
