+++ src/test/java/com/github/dakusui/scriptiveunit/unittests/preprocessing/JsonPreprocessingElementTest.java
@@ -1,6 +1,6 @@
 package com.github.dakusui.scriptiveunit.unittests.preprocessing;
 
-import com.github.dakusui.scriptiveunit.model.lang.Preprocessor;
+import com.github.dakusui.scriptiveunit.model.lang.PreprocessingElement;
 import com.github.dakusui.scriptiveunit.model.lang.HostSpec;
 import com.github.dakusui.scriptiveunit.model.lang.ApplicationSpec;
 import org.codehaus.jackson.map.ObjectMapper;
@@ -11,11 +11,11 @@ import java.io.IOException;
 
 import static org.junit.Assert.assertEquals;
 
-public class JsonPreprocessorTest {
+public class JsonPreprocessingElementTest {
   @Test
   public void whenPreprocessingOnArrayIsRequested() throws IOException {
     ObjectNode targetObject = (ObjectNode) new ObjectMapper().readTree("{\"a1\":[0,1,2]}");
-    Preprocessor jsonPreprocessor = new Preprocessor() {
+    PreprocessingElement jsonPreprocessingElement = new PreprocessingElement() {
       @Override
       public ApplicationSpec.Node translate(ApplicationSpec.Node targetElement) {
         return ApplicationSpec.dict(
@@ -26,7 +26,7 @@ public class JsonPreprocessorTest {
 
       @Override
       public boolean matches(Path pathToTargetElement) {
-        return Preprocessor.Utils.pathComponentList("a1").equals(
+        return PreprocessingElement.Utils.pathComponentList("a1").equals(
             pathToTargetElement.asComponentList()
         );
       }
@@ -34,14 +34,14 @@ public class JsonPreprocessorTest {
     HostSpec.Json hostLanguage = new HostSpec.Json();
     assertEquals(
         "{\"a1\":{\"v1\":\"Hello\",\"v2\":[0,1,2]}}",
-        hostLanguage.toHostObject(ApplicationSpec.preprocess(hostLanguage.toApplicationDictionary(targetObject), jsonPreprocessor)).toString()
+        hostLanguage.toHostObject(ApplicationSpec.preprocess(hostLanguage.toApplicationDictionary(targetObject), jsonPreprocessingElement)).toString()
     );
   }
 
   @Test
   public void whenPreprocessingOnMapIsRequested() throws IOException {
     ObjectNode targetObject = (ObjectNode) new ObjectMapper().readTree("{\"a1\":{\"c1\":100, \"c2\":200}}");
-    Preprocessor jsonPreprocessor = new Preprocessor() {
+    PreprocessingElement jsonPreprocessingElement = new PreprocessingElement() {
       @Override
       public ApplicationSpec.Node translate(ApplicationSpec.Node targetElement) {
         return ApplicationSpec.dict(
@@ -52,7 +52,7 @@ public class JsonPreprocessorTest {
 
       @Override
       public boolean matches(Path pathToTargetElement) {
-        return Preprocessor.Utils.pathComponentList("a1", "c2").equals(
+        return PreprocessingElement.Utils.pathComponentList("a1", "c2").equals(
             pathToTargetElement.asComponentList()
         );
       }
@@ -60,7 +60,7 @@ public class JsonPreprocessorTest {
     HostSpec.Json hostLanguage = new HostSpec.Json();
     assertEquals(
         "{\"a1\":{\"c1\":100,\"c2\":{\"v1\":\"Hello\",\"v2\":200}}}",
-        hostLanguage.toHostObject(ApplicationSpec.preprocess(hostLanguage.toApplicationDictionary(targetObject), jsonPreprocessor)).toString()
+        hostLanguage.toHostObject(ApplicationSpec.preprocess(hostLanguage.toApplicationDictionary(targetObject), jsonPreprocessingElement)).toString()
     );
   }
